+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## Changelog
 
+### 2026-04-06
+- DatabaseManager added: schema/seed initialization, CRUD operations for characters, inventory and coins, 5E weight calculations (some are still todo)
+- SQL files embedded in Qt resource system
+
 ### 2026-04-05
 - C++ enum definitions added (enums.h): CreatureSize, ItemType, ItemSource, ItemRarity, WeaponCategory, WeaponRangeType, DamageType, ArmorCategory